Extract tests for isUint8Array to own file
