+++ lib/referee.test.js
@@ -4,43 +4,6 @@ var referee = require("../lib/referee");
 var testHelper = require("./test-helper");
 var sinon = require("sinon");
 
-testHelper.assertionTests("assert", "isUint8Array", function(
-    pass,
-    fail,
-    msg,
-    error
-) {
-    function captureArgs() {
-        return arguments;
-    }
-
-    fail("for Uint16Array", new Uint16Array());
-    fail("for Uint32Array", new Uint32Array());
-    pass("for Uint8Array", new Uint8Array());
-    fail("for array", []);
-    fail("for object", {});
-    fail("for arguments", captureArgs());
-    msg(
-        "fail with descriptive message",
-        "[assert.isUint8Array] Expected {  } to be a Uint8Array",
-        {}
-    );
-    msg(
-        "fail with custom message",
-        "[assert.isUint8Array] Nope: Expected {  } to be a Uint8Array",
-        {},
-        "Nope"
-    );
-    error(
-        "for object",
-        {
-            code: "ERR_ASSERTION",
-            operator: "assert.isUint8Array"
-        },
-        {}
-    );
-});
-
 testHelper.assertionTests("assert", "isUint8ClampedArray", function(
     pass,
     fail,
